Fix typos and source file encoding

+++ MySqlManager.cs
@@ -3963,8 +3963,8 @@ namespace cs_elbot
                 while (reader.Read())
                 {
                     TradeHandler.SellingItem MySellingItem = new TradeHandler.SellingItem();
-                    MySellingItem.SellingItemsSqlID = reader.GetInt16(0);
-                    MySellingItem.PricelistSqlID = reader.GetInt16(1);
+                    MySellingItem.SellingItemsSqlID = reader.GetInt32(0);
+                    MySellingItem.PricelistSqlID = reader.GetInt32(1);
                     MySellingItem.KnownItemsSqlID = reader.GetInt16(2);
                     MySellingItem.announce = reader.GetBoolean(3);
                     MySellingItem.pricesale = reader.GetDouble(4);
@@ -4020,8 +4020,8 @@ namespace cs_elbot
                 while (reader.Read())
                 {
                     TradeHandler.SellingItem MySellingItem = new TradeHandler.SellingItem();
-                    MySellingItem.SellingItemsSqlID = reader.GetInt16(0);
-                    MySellingItem.PricelistSqlID = reader.GetInt16(1);
+                    MySellingItem.SellingItemsSqlID = reader.GetInt32(0);
+                    MySellingItem.PricelistSqlID = reader.GetInt32(1);
                     MySellingItem.KnownItemsSqlID = reader.GetInt16(2);
                     MySellingItem.announce = reader.GetBoolean(3);
                     MySellingItem.pricesale = reader.GetDouble(4);
@@ -4075,8 +4075,8 @@ namespace cs_elbot
                 while (reader.Read())
                 {
                     TradeHandler.WantedItem MyWantedItem = new TradeHandler.WantedItem();
-                    MyWantedItem.SellingItemsSqlID = reader.GetInt16(0);
-                    MyWantedItem.PricelistSqlID = reader.GetInt16(1);
+                    MyWantedItem.SellingItemsSqlID = reader.GetInt32(0);
+                    MyWantedItem.PricelistSqlID = reader.GetInt32(1);
                     MyWantedItem.KnownItemsSqlID = reader.GetInt16(2);
                     MyWantedItem.announce = reader.GetBoolean(3);
                     if (Settings.IsTradeBot == true)
@@ -4138,7 +4138,7 @@ namespace cs_elbot
                 while (reader.Read())
                 {
                     TradeHandler.WantedItem MyWantedItem = new TradeHandler.WantedItem();
-                    MyWantedItem.SellingItemsSqlID = reader.GetInt16(0);
+                    MyWantedItem.SellingItemsSqlID = reader.GetInt32(0);
                     WantedItemList.Add(MyWantedItem.SellingItemsSqlID);
                 }
             }
@@ -4191,8 +4191,8 @@ namespace cs_elbot
                 while (reader.Read())
                 {
                     TradeHandler.WantedItem MyWantedItem = new TradeHandler.WantedItem();
-                    MyWantedItem.SellingItemsSqlID = reader.GetInt16(0);
-                    MyWantedItem.PricelistSqlID = reader.GetInt16(1);
+                    MyWantedItem.SellingItemsSqlID = reader.GetInt32(0);
+                    MyWantedItem.PricelistSqlID = reader.GetInt32(1);
                     MyWantedItem.KnownItemsSqlID = reader.GetInt16(2);
                     MyWantedItem.announce = reader.GetBoolean(3);
                     if (Settings.IsTradeBot == true)
